csound distance & angle for one rabbit
put in the angle software bus from cpp to csound, tested it briefly. one thing i noticed is that the resonification happens at the time of touch, i think we talked about not doing it that way. a state diagram will have to be made. anyway, we have two world instruments and one rabbit instrument working but needs refined slightly. also more rabbits.